+++ NpgsqlRestClient/App.cs
@@ -252,6 +252,7 @@ public class App
         return (connection, endpoint, context) =>
         {
             var uid = context.User.FindFirstValue(options.DefaultUserIdClaimType);
+            //TODO: use ExecutionIdHeaderName from options
             var executionId = context.Request.Headers["X-Execution-Id"].FirstOrDefault();
             connection.ConnectionString = new NpgsqlConnectionStringBuilder(connectionString)
             {